requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.6.1=py36_0
astor=0.7.1=py36_0
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
ca-certificates=2019.1.23=0
cachetools=3.1.0=pypi_0
certifi=2019.3.9=py36_0
cffi=1.11.5=py36h74b6da3_1
chardet=3.0.4=pypi_0
click=7.0=py36_0
cloudpickle=0.7.0=py_0
colorama=0.4.1=py36_0
cudatoolkit=9.0=1
cudnn=7.3.1=cuda9.0_0
cycler=0.10.0=py36h009560c_0
cython=0.29.5=pypi_0
cytoolz=0.9.0.1=py36hfa6e2cd_1
dask-core=1.1.1=py_0
decorator=4.3.0=py36_0
devicehive=2.1.6=pypi_0
devicehive-webconfig=1.0.2=pypi_0
entrypoints=0.3=py36_0
ffmpeg=1.4=pypi_0
flask=1.0.2=py36_1
freetype=2.9.1=ha9979f8_1
gast=0.2.1=py36_0
geos=3.7.1=he025d50_1000
google-api-python-client=1.7.8=pypi_0
google-auth=1.6.2=pypi_0
google-auth-httplib2=0.0.3=pypi_0
grpcio=1.16.1=py36h351948d_1
h5py=2.8.0=py36hf7173ca_2
hdf5=1.8.20=hac2f561_1
httplib2=0.12.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=pypi_0
imageio=2.5.0=pypi_0
imageio-ffmpeg=0.2.0=pypi_0
imgaug=0.2.8=pypi_0
imutils=0.5.2=pypi_0
intel-openmp=2019.1=144
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.2.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36h3c5d0ee_0
ipywidgets=7.4.2=py36_0
itsdangerous=1.1.0=py36_0
jedi=0.13.2=py36_0
jinja2=2.10=py36_0
jpeg=9b=hb83a4c4_2
jsonschema=2.6.0=py36h7636477_0
jupyter=1.0.0=py36_7
jupyter_client=5.2.4=py36_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.4.0=py36_0
keras=2.2.4=0
keras-applications=1.0.6=py36_0
keras-base=2.2.4=py36_0
keras-preprocessing=1.0.5=py36_0
kiwisolver=1.0.1=py36h6538335_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.36=h2a8f88b_0
libprotobuf=3.6.1=h7bd577a_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.0.9=h36446d0_2
m2-base=1.0.0=2
m2-bash=4.3.042=5
m2-bash-completion=2.3=2
m2-catgets=1.1=3
m2-coreutils=8.25=102
m2-dash=0.5.8=2
m2-diffutils=3.3=4
m2-file=5.25=2
m2-filesystem=2016.04=4
m2-findutils=4.6.0=2
m2-gawk=4.1.3=2
m2-gcc-libs=5.3.0=4
m2-gettext=0.19.7=4
m2-gmp=6.1.0=3
m2-grep=2.22=4
m2-gzip=1.7=2
m2-inetutils=1.9.2=2
m2-info=6.0=2
m2-less=481=2
m2-libasprintf=0.19.7=4
m2-libbz2=1.0.6=3
m2-libcatgets=1.1=3
m2-libcrypt=1.3=2
m2-libgettextpo=0.19.7=4
m2-libiconv=1.14=3
m2-libintl=0.19.7=4
m2-liblzma=5.2.2=2
m2-libpcre=8.38=2
m2-libreadline=6.3.008=8
m2-libutil-linux=2.26.2=2
m2-libxml2=2.9.2=3
m2-make=4.1=5
m2-mintty=1!2.2.3=2
m2-mpfr=3.1.4=2
m2-msys2-launcher-git=0.3.28.860c495=2
m2-msys2-runtime=2.5.0.17080.65c939c=3
m2-ncurses=6.0.20160220=2
m2-sed=4.2.2=3
m2-tar=1.28=4
m2-tftp-hpa=5.2=2
m2-time=1.7=2
m2-ttyrec=1.0.8=2
m2-tzcode=2015.e=2
m2-unzip=6.0=3
m2-util-linux=2.26.2=2
m2-which=2.21=3
m2-zip=3.0=2
m2-zlib=1.2.8=4
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.0.1=py36_0
markupsafe=1.1.0=py36he774522_0
matplotlib=3.0.2=py36hc8f65d3_0
mistune=0.8.4=py36he774522_0
mkl=2019.1=144
mkl_fft=1.0.10=py36h14836fe_0
mkl_random=1.0.2=py36h343c172_0
moviepy=1.0.0=pypi_0
msys2-conda-epoch=20160418=1
nbconvert=5.3.1=py36_0
nbformat=4.4.0=py36h3a5bc1b_0
networkx=2.2=py36_1
ninja=1.8.2=py36he980bc4_1
nltk=3.4=py36_1
notebook=5.7.4=py36_0
numpy=1.15.4=py36h19fb1c0_0
numpy-base=1.15.4=py36hc3f5095_0
oauth2client=4.1.3=pypi_0
olefile=0.46=py36_0
opencv=3.4.2=py36h40b0b35_0
opencv-python=4.0.0.21=pypi_0
openssl=1.1.1b=he774522_1
pafy=0.5.4=pypi_0
pandas=0.24.1=pypi_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.3.1=py36_0
pickleshare=0.7.5=py36_0
pillow=5.4.1=py36hdc69c19_0
pip=18.1=py36_0
posix=1.0.0=2
proglog=0.1.9=pypi_0
prometheus_client=0.5.0=py36_0
prompt_toolkit=2.0.7=py36_0
protobuf=3.6.1=py36h33f27b4_0
py-opencv=3.4.2=py36hc319ecb_0
pyasn1=0.4.5=pypi_0
pyasn1-modules=0.2.4=pypi_0
pycocotools=2.0=pypi_0
pycparser=2.19=py36_0
pydrive=1.3.1=pypi_0
pydub=0.23.1=pypi_0
pygments=2.3.1=py36_0
pyparsing=2.3.0=py36_0
pyqt=5.9.2=py36h6538335_2
pyreadline=2.1=py36_1
python=3.6.8=h9f7ef89_0
python-dateutil=2.7.5=py36_0
python-wget=3.2=py_0
pytorch=1.0.1=py3.6_cuda90_cudnn7_1
pytz=2018.7=py36_0
pywavelets=1.0.1=py36h8c2d366_0
pywinpty=0.5.5=py36_1000
pyyaml=3.13=py36hfa6e2cd_0
pyzmq=17.1.2=py36ha925a31_2
qt=5.9.7=vc14h73c81de_0
qtconsole=4.4.3=py36_0
requests=2.21.0=pypi_0
rsa=4.0=pypi_0
scikit-image=0.14.2=pypi_0
scikit-learn=0.20.3=py36h343c172_0
scipy=1.1.0=py36h29ff71c_2
send2trash=1.5.0=py36_0
setuptools=40.8.0=pypi_0
shapely=1.6.4=py36h8921fb9_1003
sip=4.19.8=py36h6538335_0
six=1.12.0=py36_0
sqlite=3.26.0=he774522_0
tensorboard=1.12.2=py36h33f27b4_0
tensorflow=1.12.0=gpu_py36ha5f9131_0
tensorflow-base=1.12.0=gpu_py36h6e53903_0
tensorflow-gpu=1.12.0=h0d30ee6_0
termcolor=1.1.0=py36_1
terminado=0.8.1=py36_1
testpath=0.4.2=py36_0
tk=8.6.8=hfa6e2cd_0
toolz=0.9.0=py36_0
torchvision=0.2.1=py_2
tornado=5.1.1=py36hfa6e2cd_0
tqdm=4.28.1=py36h28b3542_0
traitlets=4.3.2=py36h096827d_0
uritemplate=3.0.0=pypi_0
urllib3=1.24.1=pypi_0
vc=14.1=h21ff451_3
vs2015_runtime=15.5.2=3
wcwidth=0.1.7=py36h3d5aa90_0
webencodings=0.5.1=py36_1
websocket-client=0.55.0=pypi_0
werkzeug=0.14.1=py36_0
wheel=0.33.1=pypi_0
widgetsnbextension=3.4.2=py36_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
yaml=0.1.7=hc54c509_2
youtube-dl=2019.3.9=pypi_0
zeromq=4.3.1=h33f27b4_3
zlib=1.2.11=h62dcd97_3
